freq commit

+++ NLPGraph/calc/levenstein_damerau.cpp
@@ -180,7 +180,7 @@ LevensteinDamerau::LevensteinDamerau(context &context)
 }
 LevensteinDamerau::~LevensteinDamerau() {
 }
-int LevensteinDamerau::calculate(uint16_t width, uint16_t haystackSize, uint64_t* needle, uint64_t* haystack, uint64_t *distancesOut, uint64_t *operationsOut) {
+int LevensteinDamerau::calculate(uint width, uint haystackSize, uint64_t* needle, uint64_t* haystack, uint64_t *distancesOut, uint64_t *operationsOut) {
 
     int result = 0;
